[refactor] Fcm 토큰 삭제시 소유자 검증 로직 추가 (#289)

+++ src/main/java/konkuk/thip/notification/application/service/FcmDeleteService.java
@@ -3,6 +3,7 @@ package konkuk.thip.notification.application.service;
 import konkuk.thip.notification.application.port.in.FcmDeleteUseCase;
 import konkuk.thip.notification.application.port.in.dto.FcmTokenDeleteCommand;
 import konkuk.thip.notification.application.port.out.FcmTokenLoadPort;
+import konkuk.thip.notification.domain.FcmToken;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -13,10 +14,12 @@ public class FcmDeleteService implements FcmDeleteUseCase {
 
     private final FcmTokenLoadPort fcmTokenLoadPort;
 
-
     @Override
     @Transactional
     public void deleteToken(FcmTokenDeleteCommand command) {
+        FcmToken fcmToken = fcmTokenLoadPort.getByDeviceIdOrThrow(command.deviceId());
+        fcmToken.validateFcmOwner(command.userId());
+
         fcmTokenLoadPort.deleteByUserIdAndDeviceId(command.userId(), command.deviceId());
     }
 }
